# swagger.yml
openapi: 3.0.3
info:
  title: Service Availability Scheduler API
  version: 0.1.0
  description: API for password login, registration, account activation, password reset, workspace administration, service availability, and reservations.
servers:
  - url: http://localhost:3000
security:
  - bearerAuth: []
...
paths:
  /api/login:
    post:
      tags: [Auth]
      security: []
      summary: Log in by email and password and receive a bearer token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated user and bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/password-reset/captcha:
    post:
      tags: [Account]
      security: []
      summary: Create a password-reset CAPTCHA challenge.
      responses:
        '200':
          description: CAPTCHA challenge prompt.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaptchaResponse'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/password-reset/request:
    post:
      tags: [Account]
      security: []
      summary: Request a password-reset token after solving CAPTCHA.
      description: Always returns ok for known and unknown emails; reset links are logged server-side until email delivery exists.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
        '400':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/Error'
  /api/password-reset/validate:
    post:
      tags: [Account]
      security: []
      summary: Validate a password-reset token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRequest'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
        '400':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/Error'
  /api/password-reset:
    post:
      tags: [Account]
      security: []
      summary: Set a new password using a valid reset token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
        '400':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/Error'
  /api/register/captcha:
    post:
      tags: [Account]
      security: []
      summary: Create a registration CAPTCHA challenge.
      responses:
        '200':
          description: CAPTCHA challenge prompt.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaptchaResponse'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/register:
    post:
      tags: [Account]
      security: []
      summary: Register a non-activated user and receive a bearer token.
      description: Activation links are logged server-side until email delivery exists. When invitation_code is supplied, the registered email must match the invitation email and member workspace access is created in the registration transaction.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Registered user and bearer token with activated=false.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/Error'
  /api/account-activation/validate:
    post:
      tags: [Account]
      security: []
      summary: Validate an account-activation token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRequest'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
        '400':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/Error'
  /api/account-activation:
    post:
      tags: [Account]
      security: []
      summary: Activate an account and receive a bearer token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRequest'
      responses:
        '200':
          description: Activated user and bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/Error'
  /api/logout:
    post:
      tags: [Auth]
      summary: Log out the current bearer-token session client-side.
      responses:
        '200':
          $ref: '#/components/responses/Ok'
        '401':
          $ref: '#/components/responses/Error'
  /api/me:
    get:
      tags: [Auth]
      summary: Return the authenticated user identity.
      responses:
        '200':
          description: Current user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Error'
  /api/renew:
    post:
      tags: [Auth]
      summary: Issue a fresh bearer token for the current user.
      responses:
        '200':
          description: Renewed token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Ok:
      description: Successful operation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/OkResponse'
    Error:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    TooManyRequests:
      description: Too many requests for the current rate-limit window.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
          example: true
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
    CaptchaResponse:
      type: object
      required: [ok, challenge_id, challenge_prompt]
      properties:
        ok:
          type: boolean
        challenge_id:
          type: string
        challenge_prompt:
          type: string
    PasswordResetRequest:
      type: object
      required: [email, challenge_id, challenge_answer]
      properties:
        email:
          type: string
          format: email
        challenge_id:
          type: string
        challenge_answer:
          type: string
        invitation_code:
          type: string
          description: Optional workspace invitation code from a valid unregistered invitation link.
    TokenRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
          description: Also accepted as reset_token or activation_token by the matching endpoint.
    ResetPasswordRequest:
      type: object
      required: [token, password, confirm_password]
      properties:
        token:
          type: string
        password:
          type: string
          format: password
          minLength: 8
        confirm_password:
          type: string
          format: password
          minLength: 8
    RegisterRequest:
      type: object
      required:
        - email
        - nickname
        - password
        - confirm_password
        - challenge_id
        - challenge_answer
      properties:
        email:
          type: string
          format: email
        nickname:
          type: string
        password:
          type: string
          format: password
          minLength: 8
        confirm_password:
          type: string
          format: password
          minLength: 8
        challenge_id:
          type: string
        challenge_answer:
          type: string
    LoginResponse:
      type: object
      required: [ok, user, token, token_type, expires_in_seconds]
      properties:
        ok:
          type: boolean
        user:
          $ref: '#/components/schemas/User'
        token:
          type: string
        token_type:
          type: string
          example: Bearer
        expires_in_seconds:
          type: integer
    User:
      type: object
      required: [id, email, nickname, activated]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
          description: Present on login responses that return the domain user object.
        email:
          type: string
          format: email
        nickname:
          type: string
        activated:
          type: boolean
...